requirements.txt
beautifulsoup4==4.4.1
bottle==0.12.9
funcsigs==0.4
matplotlib==1.4.3
mecab-python==0.996
mock==1.3.0
nose==1.3.7
numpy==1.10.1
oauthlib==1.0.3
pandas==0.17.0
pbr==1.8.1
Pillow==3.0.0
pymongo==3.0.3
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.6
requests==2.11.0
requests-oauthlib==0.5.0
scipy==0.16.0
seaborn==0.6.0
six==1.10.0
wordcloud==1.1.3
